refactor: VO 클래스들 패키지 위치 변경하기
