+++ src/main/java/lotto/controller/LottoController.java
@@ -5,16 +5,14 @@ import lotto.domain.BonusNumber;
 import lotto.domain.IssuedLottos;
 import lotto.domain.Lotto;
 import lotto.domain.Profit;
-import lotto.domain.Rank;
 import lotto.domain.LottoTicketMachine;
 import lotto.domain.WinningRanks;
 import lotto.utility.validation.WinningNumberChecker;
-import lotto.utility.vo.request.BonusNumberRequest;
-import lotto.utility.vo.response.LottoResponse;
-import lotto.utility.vo.response.ProfitResponse;
-import lotto.utility.vo.request.PurchaseAmountRequest;
-import lotto.utility.vo.request.WinningNumberRequest;
-import lotto.utility.vo.response.WinningResponse;
+import lotto.vo.request.BonusNumberRequest;
+import lotto.vo.response.LottoResponse;
+import lotto.vo.response.ProfitResponse;
+import lotto.vo.request.PurchaseAmountRequest;
+import lotto.vo.request.WinningNumberRequest;
 import lotto.view.InputView;
 import lotto.view.OutputView;
