+++ src/main/java/lotto/domain/Lotto.java
@@ -9,7 +9,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import lotto.utility.validation.LottoNumberChecker;
-import lotto.utility.vo.response.LottoResponse;
+import lotto.vo.response.LottoResponse;
 
 public class Lotto {
     private final List<Integer> numbers;
