+++ src/main/java/lotto/vo/response/WinningResponse.java
@@ -1,4 +1,4 @@
-package lotto.utility.vo.response;
+package lotto.vo.response;
 
 import java.util.ArrayList;
 import java.util.Arrays;
