+++ src/test/java/lotto/utility/vo/request/WinningNumberRequestTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import lotto.domain.Lotto;
-import lotto.utility.vo.request.WinningNumberRequest;
+import lotto.vo.request.WinningNumberRequest;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
