+++ src/test/java/lotto/utility/vo/response/WinningResponseTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Rank;
-import lotto.utility.vo.response.WinningResponse;
+import lotto.vo.response.WinningResponse;
 import org.junit.jupiter.api.Test;
 
 class WinningResponseTest {
